refactor: harden backend and clean up frontend preferences logic
- Add ALLOWED_USER_PROFILE_FIELDS whitelist to auth_preferences view for security.
- Make backend serializer factory generic based on field types.
- Extract DynamicPreferenceField to a standalone frontend component.
- Optimize UserPreferencesDialog remounting logic with stable keys.
- Fix field-multiselect grouping bug in DynamicPreferenceField.
